+++ RELEASES.md
@@ -28,6 +28,7 @@ See also GitHub releases https://github.com/lucmoreau/ProvToolbox/releases
 * prov-template-compiler: CompilerConfigurations: generate Relation0Configurator
 * prov-template-compiler: CompilerConfigurations: generate RelationConfigurator
 * prov-template-compiler: CompilerConfigurations: generate BuilderProcessorConfigurator
+* prov-template-compiler: CompilerCommon: getForeign for Builder
 
 
 `2.2.1`